Finish the UWS API definition
Add an API for listing jobs and one for waiting for a job to change
phases. Separate the latter from retrieving a job to avoid needing
special marker values for the wait parameter.

Redo how the path and operation type are specified, since using a
table results in weird formatting.